payment_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
httpx>=0.24.0
pydantic-settings>=2.0.0
